refactor: Elective Student List component
